requirements.txt
python-telegram-bot[ext]
instaloader
python-dotenv
aiofiles
gallery-dl
RedDownloader
bs4
yt-dlp
